Доделана админка, улучшен роутинг, расширены конфигурации

+++ views/admin/banners.php
@@ -6,8 +6,8 @@ require_once ROOT.'/'.Config::VIEW.'layouts/admin.php';
 <p style = "color:#1d4bee;">Здесь будут выводиться настройки баннеров</p>
 
 <?php 
-  $msg = ['sucCrt'=>'создан!','sucUpd' => 'обновлен!','sucDel' => 'удален!'];
-  $err = ['failCrt'=>'создать', 'failUpd'=>'обновить', 'failDel'=>'удалить'];
+  $msg = ['sucCrt'=>'создан!','sucUpd' => 'обновлен!','sucDel' => 'удален!', 'sucAdd'=>'сохранен!'];
+  $err = ['failCrt'=>'создать', 'failUpd'=>'обновить', 'failDel'=>'удалить', 'failAdd'=>'сохранить'];
   $style = 'display:none;';
   if (isset($_GET) && !empty($_GET)) {
     $get = array_keys($_GET)[0];
@@ -21,6 +21,26 @@ require_once ROOT.'/'.Config::VIEW.'layouts/admin.php';
   } 
 ?>
 <p style="<?=$style?>"><?=$text?></p>
+<h3>Указать статические баннеры</h3>
+<form action="<?=Config::ADDRESS?>admin/banners/add" method="post">
+  <p class="area">
+    Баннер вверху:
+    <textarea name="banner_top" class="field area" cols="30" rows="10" placeholder="<?=Config::BANNER_TOP?>"></textarea>
+  </p>
+  <p class="area">
+    Баннер слева:
+    <textarea name="banner_left" class="field area" cols="30" rows="10" placeholder="<?=Config::BANNER_LEFT?>"></textarea>
+  </p>
+  <p class="area">
+    Баннер справа:
+    <textarea name="banner_right" class="field area" cols="30" rows="10" placeholder="<?=Config::BANNER_RIGHT?>"></textarea>
+  </p>
+  <p>
+    <input type="submit" name="default" value="По умолчанию">
+    <input type="submit" name="save" value="Сохранить">
+  </p>
+</form>
+<hr/>
 
 <h3>Добавить баннер</h3>
 <form action="<?=Config::ADDRESS?>admin/banners/create" method="post">
@@ -39,18 +59,22 @@ require_once ROOT.'/'.Config::VIEW.'layouts/admin.php';
     <th>Опубликовано</th>
     <th>Удаление</th>
   </tr>
-<?php foreach ($listBanners as $item) { ?>
-  <tr>
-    <td><?=$item['title']?></td>
-    <td><?=htmlspecialchars($item['descr'])?></td>
-    <td><?=Format::adminDate($item['pubTime'])?></td>
-    <td>
-      <form action="<?=Config::ADDRESS?>admin/banners/update/<?=$item['id'];?>" method="post">
-        <?php $checked = !empty($item['publish']) ? 'checked' : '';  ?>
-        <input type="checkbox" onclick="submit();" name="publish"  <?=$checked?>>
-      </form>
-    </td>
-    <td><a href="<?=Config::ADDRESS?>admin/banners/delete/<?=$item['id']?>">Удалить</a></td>
-  </tr>
-<?php } ?>
+<?php 
+if (isset($listBanners)) {
+  foreach ($listBanners as $item) { 
+  ?>
+    <tr>
+      <td><?=$item['title']?></td>
+      <td><?=htmlspecialchars($item['descr'])?></td>
+      <td><?=Format::adminDate($item['pubTime'])?></td>
+      <td>
+        <form action="<?=Config::ADDRESS?>admin/banners/update/<?=$item['id'];?>" method="post">
+          <?php $checked = !empty($item['publish']) ? 'checked' : '';  ?>
+          <input type="checkbox" onclick="submit();" name="publish"  <?=$checked?>>
+        </form>
+      </td>
+      <td><a href="<?=Config::ADDRESS?>admin/banners/delete/<?=$item['id']?>">Удалить</a></td>
+    </tr>
+  <?php } 
+} ?>
 </table>
